Available functions

The dcoumentation of all available functions is listed below in alphabetical order. Documentation can also be found using the search bar on the left, or via the Julia REPL, e.g:

julia> using QCDNUM

julia> ?QCDNUM.evsgns

Module documentation

QCDNUM.allfxq — Method

allfxq(iset, x, qmu2, n, ichk)

Get all flavour-pdf values for a given x and mu^2.

Arguments

iset::Integer: pdf set id (1-24).

x::Float64: input value of x.

qmu2::Float64: input value of qmu2.

n::Integer: number of additional pdfs to be returned.

ichk::Integer: flag to steer error checking. See QCDNUM

docs. ichk = -1 makes code faster, at the risk of not checking certain things.

Returns

pdf::Array{Float64,2}: pdf values over flavours.

QCDNUM.asfunc — Method

asfunc(r2)

Evolve alpha_S(mu_R^2). Does not use mu^2 grid or weight tables.

Returns

alphas::FLoat64: alpha_S value.

nf::Integer: number of flavours at scale r2.

ierr::Integer: error code

QCDNUM.bvalij — Method

bvalij(iset, id, ix, iq, ichk)

Get the value of a basis pdf at a given point (ix, iq).

Arguments

iset::Integer: pdf set id (1-24)

id::Integer: basis pdf identifier from 0 to 12+n,

where n is the number of additional pdfs in iset.

ix::Integer: x index.

iq::Integer: qq index.

ichk::Integer: flag to steer error checking.

See allfxq().

Returns

pdf::Float64: pdf values.

QCDNUM.bvalxq — Method

bvalxq(iset, id, x, qq, ichk)

Get the value of a basis pdf at a given point (x, qq).

Arguments

iset::Integer: pdf set id (1-24)

id::Integer: basis pdf identifier from 0 to 12+n,

where n is the number of additional pdfs in iset.

x::Float64: x value.

qq::Float64: qq value.

ichk::Integer: flag to steer error checking.

See allfxq().

Returns

pdf::Float64: pdf values.

QCDNUM.dmpwgt — Method

dmpwgt(itype, lun, filename)

Dump weight tables of a given itype to filename.

QCDNUM.dsp_funs1 — Method

dsp_funs1(ia, u, ichk)

Evaluate function for 1D spline.

Possible values of ichk:

-1: extrapolate the spline

0: return 0

1: throw an error message

Arguments

ia::Integer: address of spline

u::Float64: x or qq

ichk::Integer: defines behaviour when outside

spline range

QCDNUM.dsp_funs2 — Method

dsp_funs2(ia, x, q, ichk)

Evaluate function for 2D spline.

Possible values of ichk:

-1: extrapolate the spline

0: return 0

1: throw an error message

Arguments

ia::Integer: address of spline

x::Float64: x value

q::Float64: qq value

ichk::Integer: defines behaviour when outside

spline range

QCDNUM.dsp_ints1 — Method

dsp_ints1(ia, u1, u2)

Evaluate integral of spline between u1 and u2.

The integration limits must lie inside the spline range.

QCDNUM.dsp_ints2 — Method

dsp_ints2(ia, x1, x2, q1, q2, rs, np)

Evaluate integral of spline between x1, x2, q1 and q2. Also takes care of rscut and integrations makes use of N-point Gauss quadrature, as defined by the choice of np.

The integration limits must lie inside the spline range.

QCDNUM.dsp_uread — Method

dsp_uread(i)

Read something from the reserved user space.

Arguments

i::Integer: where to read, from 1 to nuser

Returns

val::Float64: what is read

QCDNUM.evolfg — Method

evolfg(itype, func, def, iq0)

Evolve the flavour pdf set.

Arguments

itype::Integer: select un-polarised (1), polarised (2) or

time-like (3) evolution.

func: User-defined function that returns input x * f_j(x)

at iq0. j is from 0 to 2 * nf.

def::Array{Float64}: input array containing the contribution of

quark species i to the input distribution j.

iq0::Integer: grid index of the starting scale mu_0^2.

Returns

epsi::Float64: max deviation of the quadratic spline interpolation

from linear interpolation mid-between grid points.

QCDNUM.evsgns — Method

evsgns(itype, func, isns, n, iq0)

Evolve an arbitrary set of single/non-singlet pdfs. The evolution can only run in FFNS or MFNS mode, as it is not possible to correctly match at the thresholds as in evolfg.

Arguments

The arguments are as for evolfg, expect def::Array{Float64} is replaced with:

isns::Array{Int32,1}: Input int array specifing the

evolution type. Entries can be (+1, -1, +-2) corresponding to singlet, valence non-singlet and +/- q_ns singlets respectively.

n::Integer: Number of singlet/non-singlet pdfs to evolve

Returns

epsi::Float64: Maximum deviation of the quadratic spline from

linear interpolation mid-between the grid points.

QCDNUM.evsgnsp — Method

evsgnsp(itype, func, isns, n, iq0)

Prototype parallel version on evsgns.

QCDNUM.fillwt — Method

fillwt(itype)

Fill weight tables for all order and number of flavours. itype is used to select un-polarised pdfs (1), polarised pdfs (2) or fragmentation functions (3).

Returns

nwds::Integer: number of words used in memory.

QCDNUM.fvalij — Method

fvalij(iset, id, ix, iq, ichk)

Get the value of a flavour momentum density at a given point (ix, iq).

Arguments

iset::Integer: pdf set id (1-24)

id::Integer: basis pdf identifier from -6 to 6+n,

where n is the number of additional pdfs in iset.

ix::Integer: x index.

iq::Intger: qq index.

ichk::Integer: flag to steer error checking.

See allfxq().

Returns

pdf::Float64: pdf values.

QCDNUM.fvalxq — Method

fvalxq(iset, id, x, qq, ichk)

Get the value of a flavour momentum density at a given point (x, qq).

Arguments

iset::Integer: pdf set id (1-24)

id::Integer: basis pdf identifier from -6 to 6+n,

where n is the number of additional pdfs in iset.

x::Float64: x value.

qq::Float64: qq value.

ichk::Integer: flag to steer error checking.

See allfxq().

Returns

pdf::Float64: pdf values.

QCDNUM.gqcopy — Method

gqcopy(n)

Copy the current mu^2 grid into an array of length n.

QCDNUM.gqmake — Method

gqmake(qarr, wgt, n, nqin)

Define a logarithmically-spaced mu_F^2 grid on which the parton densities are evolved.

Arguments

qarr::Array{Float64,1}: input array containing n values of mu^2

in ascending order. The lower edge of the grid should be above 0.1 GeV^2.

wgt::Array{Float64,1}: relative grip point density in each region

defined by qarr.

n::Integer: number of values in qarr and wgt (n>=2).

nqin::Integer: requested number of grid points.

Returns

nqout::Integer: number of generated grid points.

QCDNUM.grpars — Method

grpars()

Get the current grid definitions.

Returns

nx::Integer: number of points in x grid. xmi::Float64: lower boundary of x grid. xma::Float64: upper boundary of x grid. nq::Integer: number of points in qq grid. qmi::Float64: lower boundary of qq grid. qma::Float64: upper boundary of qq grid. iord::Integer: order of spline interpolation.

QCDNUM.gxcopy — Method

gxcopy(n)

Copy the current x grid into an array of length n.

QCDNUM.gxmake — Method

gxmake(xmin, iwt, n, nxin, iord)

Define a logarithmically-spaced x grid.

Arguments

xmin::Array{Float64,1}: an input array containitn n values

of x in ascending order. xmin[1] defines the lower end of the grid and other values define approx positions where the point density will change according to the values set in iwt.

iwt::Array{Int32,1}: input integer weights given in ascending order

and must always be an integer multiple of the previous weight.

n::Integer: number of values specified in xmin and iwt.

nxin::Integer: Requested number of grid points.

iord::Integer: iord = 2(3) for linear(quadratic) spline interpolation.

Returns

nxout::Integer: the number of generated grid points.

QCDNUM.ievtyp — Method

ievtype(iset)

Get the pdf evolution type for set iset::Integer.

QCDNUM.iqfrmq — Method

iqfrmq(q2)

Get grid point index of the closest grid point at or below q2 value.

QCDNUM.isp_s2make — Method

isp_s2make(istepx, istepq)

Create a spline object in memory, and return the address. Every istep-th grid point is taken as a node point of the spline and the grid boundaries are always included as node points.

Arguments

istepx::Integer: steps taken in sampling the QCDNUM x grid

istepq::Integer: steps taken in sampling the QCDNUM qq grid

QCDNUM.isp_s2user — Method

isp_s2user(xarr, nx, qarr, nq)

Set your own node points in the spline in case the automatic sampling fails.

The routine will discard points outside the x-qq evolution grid, round the remaining nodes down to the nearest grid-point and then sort them in ascending order, discarding equal values. Thus you are allowed to enter un-sorted scattered arrays.

Arguments

xarr::Array{Float64}: array of x values

nx::Integer: length of xarr

qarr::Array{Float64}: array of qq values

nq::Integer: length of qarr

Returns

iasp::Integer: address of the spline object

QCDNUM.isp_splinetype — Method

isp_splinetype(ia)

Get the type of spline at address ia.

Possible types are:

-1: x spline

0: not a spline

1: qq spline

2: x-qq spline

QCDNUM.isp_sqmake — Method

isp_sqmake(istepq)

Create a 1D qq spline object in memory, and return the address. Every istep-th grid point is taken as a node point of the spline and the grid boundaries are always included as node points.

Arguments

istepq::Integer: steps taken in sampling the QCDNUM qq grid

QCDNUM.isp_squser — Method

isp_squser(qarr, nq)

Set your own node points in the 1D qq spline in case the automatic sampling fails.

Arguments

qarr::Array{Float64}: array of qq values

nq::Integer: length of qarr

Returns

iasp::Integer: address of the spline object

QCDNUM.isp_sxmake — Method

isp_sxmake(istepx)

Create a 1D x spline object in memory, and return the address. Every istep-th grid point is taken as a node point of the spline and the grid boundaries are always included as node points.

Arguments

istepx::Integer: steps taken in sampling the QCDNUM x grid

QCDNUM.isp_sxuser — Method

isp_sxuser(xarr, nx)

Set your own node points in the 1D x spline in case the automatic sampling fails.

Arguments

xarr::Array{Float64}: array of x values

nx::Integer: length of xarr

Returns

iasp::Integer: address of the spline object

QCDNUM.ixfrmx — Method

ixfrmx(x)

Get grid point index of closest grid point at or below x value.

QCDNUM.nptabs — Method

nptabs(iset)

Get the number of pdf tables in set iset::Integer.

QCDNUM.nxtlun — Method

nxtlun(lmin)

Get next free logical unit number above max(lmin, 10). Returns 0 if there is no free logical unit.

QCDNUM.qcinit — Method

qcinit(lun, filename)

Initialise QCDNUM - should be called before anything else.

Arguments

lun::Integer: the output logical unit number. When set to 6,

the QCDNUM messages appear on the standard output. When set to -6, the QCDNUM banner printout is suppressed.

filename::String: the output filename to store log.

QCDNUM.qfrmiq — Method

qfrmiq(iq)

Get q2 value at grid point iq.

QCDNUM.qstore — Method

qstore(action, i, val)

QCDNUM reserves 500 words of memory for user use.

Arguments

action::String: Can be "write", "read", "lock" or "unlock".

i::Integer: Where to read/write in the store.

val::Float64: What to write in the store.

QCDNUM.qstore — Method

qstore(action, i)

QCDNUM reserves 500 words of memory for user use.

Arguments

action::String: Can be "write", "read", "lock" or "unlock".

i::Integer: Where to read/write in the store.

Returns

val::Float64: What is read from the store.

QCDNUM.readwt — Method

readwt(lun, filename)

Read weight tables from filename. TODO: Fix implementation which currently crashes.

QCDNUM.setabr — Method

setabr(ar, br)

Define the relation between the factorisation scale muF^2 and the renormalisation scale muR^2.

muR^2 = aR muF^2 + bR.

QCDNUM.setalf — Method

setalf(alfs, r2)

Set the starting value of alpha_S and the starting renormalisation scale r2. By default alpha_S(m_Z^2) = 0.118.

QCDNUM.setcbt — Method

setcbt(nfix, iqc, iqb, iqt)

Select FFNS or VFNS, and thresholds on mu_F^2 if necessary.

Arguments

nfix::Integer: number of flavours in the FFNS. For VNFS set

nfix = 0.

iqc/b/t::Integer: grid indices of the heavy flavour thresholds

in the VFNS. Ignored if FFNS.

QCDNUM.setint — Method

setint(param, val)

Set or get QCDNUM integer parameters.

Arguments

param::String: Name of parameter. Can be "iter" (number of

evolutions in backwards iteration), "tlmc" (time-like matching conditions), "nopt" (number of perturbative terms) or "edbg" (evolution loop debug printout).

val::Integer: Value to set.

QCDNUM.setord — Method

setord(iord)

Set order of perturbative QCD calculations. iord = 1, 2, 3 for LO, NLO and NNLO respectively. By default, iord = 2.

QCDNUM.ssp_erase — Method

ssp_erase(ia)

Clear the memory from ia onwards. ia=0 can also be used to erase all spline objects in memory.

QCDNUM.ssp_extrapu — Method

ssp_extrapu(ia, n)

Define the extrapolation at the kinematic limit for a spline at address ia.

The extrapolation index n can be:

0: constant

1: linear

2: quadratic

3: cubic

QCDNUM.ssp_extrapv — Method

ssp_extrapv(ia, n)

Define the extrapolation at the kinematic limit for a spline at address ia.

The extrapolation index n can be:

0: constant

1: linear

2: quadratic

3: cubic

QCDNUM.ssp_nprint — Method

ssp_nprint(ia)

Print a list of nodes and grids for the spline at address ia.

QCDNUM.ssp_s2f123 — Method

ssp_s2f123(ia, iset, def, istf, rs)

Fast structure function input for 2D splines over x and qq.

Arguments

ia::Integer: address of the spline object

iset::Integer: QCDNUM pdf-set index

def::Array{Float64}: Array of (anti-)quark

coefficients

istf::Integer: structure function index

1<=>FL, 2<=>F2, 3<=>xF3, 4<=>FL'

rs::Float64: sqrt(s) cut - 0 for no kinematic cut

QCDNUM.ssp_s2fill — Method

ssp_s2fill(iasp, fun, rs)

Fill the spline object by passing a function. The function must have the signature fun(ix::Integer, iq::Integer, first::Boolean).

Arguments

iasp::Integer: address of the spline object

fun::Ptr{Nothing}: function to be splined

rs::Float64: set a sqrt(s) cut - 0 for no kinematic cut

QCDNUM.ssp_spinit — Method

ssp_spinit(nuser)

Initialise SPLINT - should be called before other SPLINT functions.

Arguments

nuser::Integer: the number of words reserved for user storage

QCDNUM.ssp_splims — Method

ssp_splims(ia)

Get node limits of spline at address ia.

Here, u and v refer to the x and qq dimensions respectively.

Returns tuple containing

nu::Integer: number of nodes in u direction

u1::Float64: lower u limit

u2::Float64: upper u limit

nv::Integer: number of nodes in v direction

v1::Float64: lower v limit

v2::Float64: upper v limit

n::Integer: number of active nodes below kinematic limit

QCDNUM.ssp_sqf123 — Method

ssp_sq123(ia, iset, def, istf, ix)

Fast structure function input for splines over qq.

Arguments

ia::Integer: address of the spline object

iset::Integer: QCDNUM pdf-set index

def::Array{Float64}: Array of (anti-)quark

coefficients

istf::Integer: structure function index

1<=>FL, 2<=>F2, 3<=>xF3, 4<=>FL'

ix::Integer: index of x value

QCDNUM.ssp_sqfill — Method

ssp_sqfill(iasp, fun, ix)

Fill the 1D qq spline object by passing a function. The function must have the signature fun(ix::Integer, iq::Integer, first::Boolean).

Arguments

iasp::Integer: address of the spline object

fun::Ptr{Nothing}: function to be splined

ix::Integer: fixed ix value to pass to fun

QCDNUM.ssp_sxf123 — Method

ssp_sxf123(ia, iset, def, istf, iq)

Fast structure function input for splines over x.

Arguments

ia::Integer: address of the spline object

iset::Integer: QCDNUM pdf-set index

def::Array{Float64}: Array of (anti-)quark

coefficients

istf::Integer: structure function index

1<=>FL, 2<=>F2, 3<=>xF3, 4<=>FL'

iq::Integer: index of qq value

QCDNUM.ssp_sxfill — Method

ssp_sxfill(iasp, fun, iq)

Fill the 1D x spline object by passing a function. The function must have the signature fun(ix::Integer, iq::Integer, first::Boolean).

Arguments

iasp::Integer: address of the spline object

fun::Ptr{Nothing}: function to be splined

iq::Integer: fixed iq value to pass to fun

QCDNUM.ssp_unodes — Method

ssp_unodes(ia, n, nu)

Copy u-nodes from spline at address ia to local array.

Arguments

ia::Integer: address of spline

n::Integer: dimension of array to copy to

nu::Integer: number of u-nodes copied

Returns

array::Array{Float64}: array of u-nodes

QCDNUM.ssp_uwrite — Method

ssp_uwrite(i, val)

Write something to the reserved user space.

Arguments

i::Integer: where to write, from 1 to nuser

val::Float: what to write

QCDNUM.ssp_vnodes — Method

ssp_vnodes(ia, n, nv)

Copy v-nodes from spline at address ia to local array.

Arguments

ia::Integer: address of spline

n::Integer: dimension of array to copy to

nv::Integer: number of v-nodes copied

Returns

array::Array{Float64}: array of v-nodes

QCDNUM.sumfxq — Method

sumfxq(iset, c, isel, x, qmu2, ichk)

Return the gluon or a weighted sum of quark densities, depending on the selection flag, isel.

isel values:

0: Gluon density |xg> 1: Linear combination c summed over active flavours 2-8: Specific singlet/non-singlet quark component 9: Intrinsic heavy flavours 12+i: Additional pdf |xf_i> in iset

See QCDNUM manual for more information.

Arguments

iset::Integer: pdf set id (1-24).

c::Array{Float64}: Coefficients of quarks/anti-quarks

isel::Integer: Selection flag

x::Float64: input value of x.

qmu2::Float64: input value of qmu2.

n::Integer: number of additional pdfs to be returned.

ichk::Integer: flag to steer error checking. See QCDNUM

docs. ichk = -1 makes code faster, at the risk of not checking certain things.

Returns

pdf::Float64: pdf value.

QCDNUM.usrpdf — Method

usrpdf(fun, iset, n, offset)

Create a user-defined type-5 pdfset (same type as the output of evsgns).

Arguments

fun: User-defined function with the signature

fun(ipdf::Integer, x::Float64, qq::Float64, first::UInt8)::Float64 specifying the values at x and qq of pdfset ipdf.

iset::Integer: Pdfset identifier, between 1 and 24.

n::Integer: Number of pdf tables in addition to gluon tables.

offset::Float64: Relative offset at the thresholds mu_h^2, used

to catch matching discontinuities.

Returns

epsi::Float64: Maximum deviation of the quadratic spline from

linear interpolation mid-between the grid points.

QCDNUM.wtfile — Method

wtfile(itype, filename)

Maintains an up-to-date weight table in filename.

QCDNUM.xfrmix — Method

xfrmix(ix)

Get x value at grid point ix.

QCDNUM.zmfillw — Method

zmfillw()

Fill weight tables for zero-mass structure function calculations.

QCDNUM.zmstfun — Method

zmstfun()

Calculate a structure function from a linear combination of parton densities.

Arguments

istf::Integer: structure function index

where (1,2,3,4) = (FL, F2, xF3, fL^').

def::Array{Float64}: coeffs of the quark linear combination

for which the structure function is to be calculated.

x::Array{Float64}: list of x values.

Q2::Array{Float64}: list of Q2 values.

n::Integer: number of items in x, Q2 and f.

ichk::Integer: flag for grid boundary checks. See QCDNUM docs.

Returns

f::Array{Float64}: list of structure functions.

QCDNUM.zmwords — Method

zmwords()

Check the number of words available in the ZMSTF workspace and the number of words used.